perf:arm64: fix pmu percpu irq handling at hotplug.
If the cpu pmu is using a percpu irq:
1. When a cpu is down, we should disable pmu irq on
that cpu. Otherwise, if the cpu is still down when
the last perf event is released, the pmu irq can't
be freed. because the irq is still enabled on the
offlined cpu. And following perf_event_open()
syscalls will fail.

2. When a cpu is up, we should enable pmu irq on
that cpu. Otherwise, profiling tools can't sample
events on that cpu before all perf events are
released, because pmu irq is disabled on that cpu.

Bug: 30971326

Change-Id: I912db0ac9ba764582d0130c815b15399fe77c8f6
(cherry picked from commit 71b352439dfa9e3b593d2bb4ca8400c1b213ffab)